requirements.txt
aiogram==2.25.1
gspread==5.10.0
google-auth==2.24.0
